+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.13.0
+## 0.13.0 - 2022-05-31
 
 * Contributes Startback::Model, Startback::Services, as well
   as Startback::Support::DataObject and Startback::Support::World